data manager class refactoring
