+++ src/main/java/online/talkandtravel/exception/auth/AuthenticationException.java
@@ -1,4 +1,6 @@
-package online.talkandtravel.exception;
+package online.talkandtravel.exception.auth;
+
+import online.talkandtravel.exception.model.ApiException;
 
 public class AuthenticationException extends ApiException {
