+++ src/main/java/online/talkandtravel/exception/data/FailedToReadJsonException.java
@@ -1,0 +1,10 @@
+package online.talkandtravel.exception.data;
+
+public class FailedToReadJsonException extends RuntimeException {
+
+  private static final String MESSAGE = "Failed to parse JSON file at %s";
+
+  public FailedToReadJsonException(String jsonFilePath) {
+    super(MESSAGE.formatted(jsonFilePath));
+  }
+}
