+++ src/main/java/online/talkandtravel/service/CountryService.java
@@ -8,9 +8,6 @@ public interface CountryService {
 
     List<CountryInfoDto> getAllCountriesInfo();
 
-
-    void createInitialChats();
-
     /*boolean userIsSubscribed(String countryName, Long userId);
 
     Country createAndSave(Country country);
