+++ src/main/java/online/talkandtravel/service/impl/AuthenticationServiceImpl.java
@@ -1,8 +1,8 @@
 package online.talkandtravel.service.impl;
 
-import online.talkandtravel.exception.AuthenticationException;
-import online.talkandtravel.exception.RegistrationException;
-import online.talkandtravel.exception.UserNotFoundException;
+import online.talkandtravel.exception.auth.AuthenticationException;
+import online.talkandtravel.exception.auth.RegistrationException;
+import online.talkandtravel.exception.user.UserNotFoundException;
 import online.talkandtravel.model.entity.Role;
 import online.talkandtravel.model.entity.Token;
 import online.talkandtravel.model.entity.TokenType;
