+++ src/main/java/online/talkandtravel/service/impl/AvatarServiceImpl.java
@@ -1,8 +1,8 @@
 package online.talkandtravel.service.impl;
 
-import online.talkandtravel.exception.FileSizeExceededException;
-import online.talkandtravel.exception.ImageProcessingException;
-import online.talkandtravel.exception.UnsupportedFormatException;
+import online.talkandtravel.exception.file.FileSizeExceededException;
+import online.talkandtravel.exception.file.ImageProcessingException;
+import online.talkandtravel.exception.file.UnsupportedFormatException;
 import online.talkandtravel.model.entity.Avatar;
 import online.talkandtravel.repository.AvatarRepository;
 import online.talkandtravel.service.AvatarService;
