+++ src/main/java/online/talkandtravel/util/dataManager/DataManager.java
@@ -1,0 +1,13 @@
+package online.talkandtravel.util.dataManager;
+
+import java.io.IOException;
+import org.springframework.transaction.annotation.Transactional;
+
+public interface DataManager {
+
+  void checkAndPopulateCountries();
+
+  void checkAndCreateChats();
+
+  void addAdmin();
+}
